Joined GLSLPostfixExpression with GLSLPrefixOperatorExpression into one GLSLUnaryOperatorExpression
